+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+### v1.7.0 (2020-11-16)
+
 * Add DateTimeImmutableFactory and DateString methods for dealing with microsecond-precision 
   date/time values
 * Improve test assertion API on the MockMutexWrapper